requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
anthropic>=0.7.0
beautifulsoup4>=4.12.0
textblob>=0.17.0
feedparser>=6.0.0
scikit-learn>=1.3.0
pytz>=2021.1